+++ src/test/kotlin/me/bossm0n5t3r/leetcode/coinChange/CoinChangeTest.kt
@@ -1,10 +1,10 @@
-package coinChange
+package me.bossm0n5t3r.leetcode.coinChange
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 internal class CoinChangeTest {
-    private val coinChange = CoinChange()
+    private val coinChange = CoinChange.Solution()
 
     private data class CoinChangeTestData(
         val coins: IntArray,
